feat(context): compartilhando estado dos cafés
